requirements.txt
flask
PyJWT
mysql-connector-python
#-e .